feat: Add RandomJokesScreen for displaying random jokes with error handling
