--- bricks/flutter_starter_brick/__brick__/lib/features/random_jokes/random_jokes_screen/random_jokes_screen.dart
@@ -1,2 +1,0 @@
-export 'src/blocs/jokes_cubit/jokes_cubit.dart';
-export 'src/ui/random_jokes_screen.dart';
